Move component into UI folder
