+++ src/components/cluster/detail/cluster_detail_table.js
@@ -1,6 +1,6 @@
 import { relativeDate } from '../../../lib/helpers.js';
 import _ from 'underscore';
-import AvailabilityZonesLabel from '../../shared/availability_zones_label';
+import AvailabilityZonesLabel from '../../UI/availability_zones_label';
 import AWSAccountID from '../../UI/aws_account_id';
 import Button from '../../UI/button';
 import moment from 'moment';
